// src/types/permission.ts
export interface Permission {
  id: number;
  resource: string;
  action: string;
  description: string;
  createdAt: string;
  updatedAt: string;
}

export enum PermissionEnum {
  PROBLEM_CREATE = 1,
  PROBLEM_READ = 2,
  PROBLEM_UPDATE = 3,
  PROBLEM_DELETE = 4,
  PROBLEM_READ_ALL = 5,
  SUBMISSION_CREATE = 6,
  SUBMISSION_READ = 7,
  SUBMISSION_READ_ALL = 8,
  CONTEST_CREATE = 9,
  CONTEST_READ = 10,
  CONTEST_UPDATE = 11,
  CONTEST_DELETE = 12,
  CONTEST_JOIN = 13,
  CONTEST_JOIN_PREMIUM = 14,
  USER_READ = 15,
  USER_UPDATE = 16,
  USER_BAN = 17,
  USER_DELETE = 18,
  ADMIN_ACCESS = 19,
  ADMIN_ROLES = 20,
  ADMIN_STATISTICS = 21,
  AI_INTERVIEW = 22,
  AI_HINT = 23,
  AI_UNLIMITED = 24,
  POST_CREATE = 25,
  POST_UPDATE = 26,
  POST_DELETE = 27,
  COMMENT_CREATE = 28,
  COMMENT_DELETE = 29,
  PAYMENT_CREATE = 30,
  LANGUAGE_CREATE = 31,
  LANGUAGE_READ = 32,
  LANGUAGE_UPDATE = 33,
  LANGUAGE_DELETE = 34,
  LANGUAGE_ACTIVATE = 35,
  LANGUAGE_DEACTIVATE = 36,
  COMMENT_READ = 37,
  COMMENT_UPDATE = 38,
  COMMENT_REPORT_READ = 39,
  COMMENT_REPORT_UPDATE = 40,
  SAMPLE_TESTCASE_CREATE = 41,
  SAMPLE_TESTCASE_READ = 42,
  SAMPLE_TESTCASE_UPDATE = 43,
  SAMPLE_TESTCASE_DELETE = 44,
  SOLUTION_CREATE = 45,
  SOLUTION_READ = 46,
  SOLUTION_UPDATE = 47,
  SOLUTION_DELETE = 48,
  SOLUTION_VOTE = 49,
  SOLUTION_COMMENT_CREATE = 50,
  SOLUTION_COMMENT_READ = 51,
  SOLUTION_COMMENT_UPDATE = 52,
  SOLUTION_COMMENT_DELETE = 53,
  SOLUTION_COMMENT_VOTE = 54,
  ROLE_CREATE = 55,
  ROLE_READ = 56,
  ROLE_UPDATE = 57,
  ROLE_DELETE = 58,
}